[update] 変数の文字列展開(add_data.php)、エラーページへ遷移するコード追加(delete01.php)
・add_data.php
SQL文にGETしたものを追加する形で実行するように変更しました。

・delete01.php
$idを取得できなかった際のエラー、DBへアクセスできなかった時のエラーを追加しました。

+++ system/delete01.php
@@ -1,13 +1,21 @@
 <?php
     include("../include/DB_access.php");
-    $id = $_GET['member_ID'];
-    $delete_str="DELETE FROM member WHERE member_ID='$id'";
-    //本番ではコメントアウト外す
 
-    $sql = $pdo->prepare($delete_str);
-    $sql->execute();
-    
-    #echo $delete_str;
-    //トップ画面へ遷移
-    header('Location:index.php');
+    $id = $_GET['member_ID'];
+    if(isset($id) && !empty($id)){
+        $delete_str="DELETE FROM member WHERE member_ID='$id'";
+
+        try{
+            $sql = $pdo->prepare($delete_str);
+            $sql->execute();
+
+            //トップ画面へ遷移
+            header('Location:index.php');
+
+        }catch(PDOException $e){
+            header('Location:../include/error.php');
+        }
+    }else{
+        header('Location:../include/error.php');
+    }
 ?>
